+++ Examples/Resources/SqlRS/3-CustomConfigurationUsingSsl.ps1
@@ -2,15 +2,10 @@
 .EXAMPLE
     This example performs a custom SQL Server Reporting Services configuration.
     It will initialize SQL Server Reporting Services and register the below
-    custom Report Server Web Service and Report Manager URLs.
+    custom Report Server Web Service and Report Manager URLs and enable SSL.
 
-    Report Server Web Service:
-    http://localhost:80/MyReportServer
-    https://localhost:443/MyReportServer
-
-    Report Manager:
-    http://localhost:80/MyReports
-    https://localhost:443/MyReports
+    Report Server Web Service: https://localhost:443/MyReportServer ()
+    Report Manager: https://localhost:443/MyReports
 
     Note: this resource does not currently handle SSL bindings for HTTPS endpoints.
 #>
@@ -26,8 +21,9 @@ Configuration Example
             DatabaseInstanceName         = 'MSSQLSERVER'
             ReportServerVirtualDirectory = 'MyReportServer'
             ReportsVirtualDirectory      = 'MyReports'
-            ReportServerReservedUrl      = @('http://+:80', 'https://+:443')
-            ReportsReservedUrl           = @('http://+:80', 'https://+:443')
+            ReportServerReservedUrl      = @('https://+:443')
+            ReportsReservedUrl           = @('https://+:443')
+            UseSsl                       = $true
         }
     }
 }