+++ Tests/Integration/MSFT_SqlRS.Integration.Tests.ps1
@@ -29,10 +29,6 @@ $mockSqlInstallAccountPassword = ConvertTo-SecureString -String 'P@ssw0rd1' -AsP
 $mockSqlInstallAccountUserName = "$env:COMPUTERNAME\SqlInstall"
 $mockSqlInstallCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $mockSqlInstallAccountUserName, $mockSqlInstallAccountPassword
 
-$mockSqlAdminAccountPassword = ConvertTo-SecureString -String 'P@ssw0rd1' -AsPlainText -Force
-$mockSqlAdminAccountUserName = "$env:COMPUTERNAME\SqlAdmin"
-$mockSqlAdminCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $mockSqlAdminAccountUserName, $mockSqlAdminAccountPassword
-
 $mockReportingServicesServiceAccountPassword = ConvertTo-SecureString -String 'yig-C^Equ3' -AsPlainText -Force
 $mockReportingServicesServiceAccountUserName = "$env:COMPUTERNAME\svc-Reporting"
 $mockReportingServicesServiceCredential = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $mockReportingServicesServiceAccountUserName, $mockReportingServicesServiceAccountPassword
@@ -58,7 +54,6 @@ try
                 {
                     $configurationParameters = @{
                         SqlInstallCredential               = $mockSqlInstallCredential
-                        SqlAdministratorCredential         = $mockSqlAdminCredential
                         ReportingServicesServiceCredential = $mockReportingServicesServiceCredential
                         OutputPath                         = $TestDrive
                         # The variable $ConfigurationData was dot-sourced above.
@@ -97,6 +92,7 @@ try
                 $resourceCurrentState.DatabaseServerName | Should -Be $mockDatabaseServerName
                 $resourceCurrentState.DatabaseInstanceName | Should -Be $mockDatabaseInstanceName
                 $resourceCurrentState.IsInitialized | Should -Be $true
+                $resourceCurrentState.UseSsl | Should -Be $false
             }
 
             It 'Should be able to access the ReportServer site without any error' {
@@ -127,6 +123,120 @@ try
                 try
                 {
                     $webRequestReportServer = Invoke-WebRequest -Uri $reportsUri -UseDefaultCredentials
+                    # if the request finishes successfully this should return status code 200.
+                    $webRequestStatusCode = $webRequestReportServer.StatusCode -as [int]
+                }
+                catch
+                {
+                    <#
+                        If the request generated an exception i.e. "HTTP Error 503. The service is unavailable."
+                        we can pull the status code from the Exception.Response property.
+                    #>
+                    $webRequestResponse = $_.Exception.Response
+                    $webRequestStatusCode = $webRequestResponse.StatusCode -as [int]
+                }
+
+                $webRequestStatusCode | Should -BeExactly 200
+            }
+        }
+
+        $configurationName = "$($script:DSCResourceName)_InstallReportingServices_ConfigureSsl_Config"
+
+        Context ('When using configuration {0}' -f $configurationName) {
+            It 'Should compile and apply the MOF without throwing' {
+                {
+                    $configurationParameters = @{
+                        SqlInstallCredential               = $mockSqlInstallCredential
+                        OutputPath                         = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData                  = $ConfigurationData
+                    }
+
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path         = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait         = $true
+                        Verbose      = $true
+                        Force        = $true
+                        ErrorAction  = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
+                } | Should -Not -Throw
+            }
+
+            It 'Should have set the resource and all the parameters should match' {
+                $currentConfiguration = Get-DscConfiguration
+
+                $resourceCurrentState = $currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName
+                } | Where-Object -FilterScript {
+                    $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.UseSsl | Should -Be $true
+            }
+
+            <#
+                We expect this to throw any error. Usually 'Unable to connect to the remote server' but it
+                can also throw and 'The underlying connection was closed: An unexpected error occurred on a send'.
+                When we support SSL fully with this resource, this should not throw at all. So leaving this
+                as this without testing for the correct error message on purpose.
+            #>
+            It 'Should not be able to access the ReportServer site and throw an error message' {
+                $reportServerUri = 'http://{0}/ReportServer_{1}' -f $env:COMPUTERNAME, $mockInstanceName
+
+                { Invoke-WebRequest -Uri $reportServerUri -UseDefaultCredentials } | Should -Throw
+            }
+        }
+
+        $configurationName = "$($script:DSCResourceName)_InstallReportingServices_RestoreToNoSsl_Config"
+
+        Context ('When using configuration {0}' -f $configurationName) {
+            It 'Should compile and apply the MOF without throwing' {
+                {
+                    $configurationParameters = @{
+                        SqlInstallCredential               = $mockSqlInstallCredential
+                        OutputPath                         = $TestDrive
+                        # The variable $ConfigurationData was dot-sourced above.
+                        ConfigurationData                  = $ConfigurationData
+                    }
+
+                    & $configurationName @configurationParameters
+
+                    $startDscConfigurationParameters = @{
+                        Path         = $TestDrive
+                        ComputerName = 'localhost'
+                        Wait         = $true
+                        Verbose      = $true
+                        Force        = $true
+                        ErrorAction  = 'Stop'
+                    }
+
+                    Start-DscConfiguration @startDscConfigurationParameters
+                } | Should -Not -Throw
+            }
+
+            It 'Should have set the resource and all the parameters should match' {
+                $currentConfiguration = Get-DscConfiguration
+
+                $resourceCurrentState = $currentConfiguration | Where-Object -FilterScript {
+                    $_.ConfigurationName -eq $configurationName
+                } | Where-Object -FilterScript {
+                    $_.ResourceId -eq $resourceId
+                }
+
+                $resourceCurrentState.UseSsl | Should -Be $false
+            }
+
+            It 'Should be able to access the ReportServer site without any error' {
+                $reportServerUri = 'http://{0}/ReportServer_{1}' -f $env:COMPUTERNAME, $mockInstanceName
+
+                try
+                {
+                    $webRequestReportServer = Invoke-WebRequest -Uri $reportServerUri -UseDefaultCredentials
                     # if the request finishes successfully this should return status code 200.
                     $webRequestStatusCode = $webRequestReportServer.StatusCode -as [int]
                 }