+++ Tests/Integration/MSFT_SqlRS.config.ps1
@@ -42,12 +42,7 @@ Configuration MSFT_SqlRS_InstallReportingServices_Config
         [Parameter(Mandatory = $true)]
         [ValidateNotNullOrEmpty()]
         [System.Management.Automation.PSCredential]
-        $ReportingServicesServiceCredential,
-
-        [Parameter(Mandatory = $true)]
-        [ValidateNotNullOrEmpty()]
-        [System.Management.Automation.PSCredential]
-        $SqlAdministratorCredential
+        $ReportingServicesServiceCredential
     )
 
     Import-DscResource -ModuleName 'PSDscResources'
@@ -95,11 +90,6 @@ Configuration MSFT_SqlRS_InstallReportingServices_Config
             SuppressReboot        = $Node.SuppressReboot
             ForceReboot           = $Node.ForceReboot
 
-            # This must be set if using SYSTEM account to install.
-            SQLSysAdminAccounts   = @(
-                $SqlAdministratorCredential.UserName
-            )
-
             DependsOn             = @(
                 '[xMountImage]MountIsoMedia'
                 '[User]CreateReportingServicesServiceAccount'
@@ -126,6 +116,68 @@ Configuration MSFT_SqlRS_InstallReportingServices_Config
             DependsOn            = @(
                 '[SqlSetup]InstallReportingServicesInstance'
             )
+        }
+    }
+}
+
+Configuration MSFT_SqlRS_InstallReportingServices_ConfigureSsl_Config
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.Management.Automation.PSCredential]
+        $SqlInstallCredential
+    )
+
+    Import-DscResource -ModuleName 'SqlServerDsc'
+
+    node localhost {
+        SqlRS 'Integration_Test'
+        {
+            # Instance name for the Reporting Services.
+            InstanceName         = $Node.InstanceName
+            UseSsl               = $true
+
+            <#
+                Instance for Reporting Services databases.
+                Note: This instance is created in a prior integration test.
+            #>
+            DatabaseServerName   = $Node.DatabaseServerName
+            DatabaseInstanceName = $Node.DatabaseInstanceName
+
+            PsDscRunAsCredential = $SqlInstallCredential
+        }
+    }
+}
+
+Configuration MSFT_SqlRS_InstallReportingServices_RestoreToNoSsl_Config
+{
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [System.Management.Automation.PSCredential]
+        $SqlInstallCredential
+    )
+
+    Import-DscResource -ModuleName 'SqlServerDsc'
+
+    node localhost {
+        SqlRS 'Integration_Test'
+        {
+            # Instance name for the Reporting Services.
+            InstanceName         = $Node.InstanceName
+            UseSsl               = $false
+
+            <#
+                Instance for Reporting Services databases.
+                Note: This instance is created in a prior integration test.
+            #>
+            DatabaseServerName   = $Node.DatabaseServerName
+            DatabaseInstanceName = $Node.DatabaseInstanceName
+
+            PsDscRunAsCredential = $SqlInstallCredential
         }
     }
 }